requirements.txt
flask==1.1.2
gunicorn==20.0.4
python-docx==0.8.11
scikit-learn==0.22.1
catboost==0.26.1
numpy==1.19.5
pandas==1.1.5
nltk==3.6.2